Reorganize and clean up twist_controller package.
- Remove file extensions for executables
- Move executables to scripts directory
- Move modules into a package and add catkin build instructions for Python package
- Clean up CMakeLists.txt
- Remove unused dependency on styx_msgs
- Clean up packages.xml
