+++ ros/src/twist_controller/launch/dbw.launch
@@ -1,6 +1,6 @@
 <?xml version="1.0"?>
 <launch>
-    <node pkg="twist_controller" type="dbw_node.py" name="dbw_node">
+    <node pkg="twist_controller" type="dbw_node" name="dbw_node">
         <param name="vehicle_mass" value="1736.35" />
         <param name="fuel_capacity" value="13.5" />
         <param name="brake_deadband" value=".1" />
