+++ ros/src/twist_controller/launch/dbw_sim.launch
@@ -1,6 +1,6 @@
 <?xml version="1.0"?>
 <launch>
-    <node pkg="twist_controller" type="dbw_node.py" name="dbw_node" output="screen">
+    <node pkg="twist_controller" type="dbw_node" name="dbw_node" output="screen">
         <param name="vehicle_mass" value="1080." />
         <param name="fuel_capacity" value="0." />
         <param name="brake_deadband" value="1." />
