+++ ros/src/twist_controller/setup.py
@@ -1,0 +1,13 @@
+## ! DO NOT MANUALLY INVOKE THIS setup.py, USE CATKIN INSTEAD
+
+from distutils.core import setup
+from catkin_pkg.python_setup import generate_distutils_setup
+
+# fetch values from package.xml
+setup_args = generate_distutils_setup(
+    packages=['twist_controller'],
+    package_dir={'': 'src'},
+    requires=['rospy']
+)
+
+setup(**setup_args)
